Improved test example
